pyproject.toml
[tool.poetry]
name = "feishu-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pydantic = "^2.4.2"
lark-oapi = "^1.0.10"
apscheduler = "^3.10.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
